test(embed): extract + unit-test buildEmbedChartOption to lock the #1103 byte-identical guarantee
Fast-follow to #1364. The reader (embedChartTheme.ts) was unit-tested, but the
APPLIER — buildOption inside EmbedChart.svelte (the CSS-var -> ECharts
palette/axis mapping + the "byte-identical when unstyled" back-compat guarantee)
— had no direct test (correct-by-construction only, so a future buildOption
refactor could silently break it).

- Extract buildOption (+ isNumericColumn) VERBATIM from EmbedChart.svelte into a
  pure embedChartOption.ts (exported buildEmbedChartOption); the component now
  imports it. Behaviour-preserving — the existing 1015 vitest tests still pass.
- Add embedChartOption.test.ts (8 tests): unstyled emits NO color/fg/axisLine
  keys (the byte-identical lock), palette -> color cycle, fg -> axis+legend text,
  axisLine -> axis lines + y splitLine, bar/line/pie series shaping, empty-rows
  "No data".

svelte-check 0 errors, vitest 1023/1023 (1015 + 8), embed bundle builds.
